MINOR: cpu-topo: assign capacity based on thread count if missing
If CPU capacities are not set based on nominal_perf, scaling_max_freq,
but we figure some cores are SMT-capable and others not, then we'll
assign 100 to the SMT-capable ones ("performance cores") and 50 to the
other ones ("efficient cores"). This can be used as a workaround on
systems where no ACPI nor cpu_freq is exposed.